requirements.txt
fastapi
uvicorn
pydantic
requests
beautifulsoup4
PyGithub
openai
python-dotenv